Ensure writes to recording files are atomic
For small request/response pairs this isn't necessary and the locks will
probably slow us down a bit, as the major OSes all guarantee atomicity
of small writes.

For larger requests, this ensures that we do not corrupt the recordings
by concurrent writes to the same file, allowing us to run multiple tests
with VcrMode::Record to the same file.