Simplify recording unit test
The second client is not under test, but verifies the behavior of the
VCR being tested; therefore we can and should be explicit about our
expectations rather than use the middleware to modify
requests/responses.

Our cookie assertion is potentially problematic; we have no real control
over which cookie is de/serialized first, so no control over which is
removed. If we re-record this test it could fail. It might be better to
just assert that the vector's length is 1.